pyproject.toml
[project]
name = "fastai"
version = "0.0.1"
description = "FastAI Repository for backend developers."
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "aioboto3>=15.1.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "editorconfig-checker==3.2.1",
    "fastapi[standard]==0.115.12",
    "gotenberg-api",
    "html-page-generator",
    "mypy>=1.17.0",
    "pydantic-settings>=2.10.1",
    "ruff==0.11.13",
]

[tool.uv.sources]
html-page-generator = { git = "https://github.com/devmanorg/html-page-generator.git" }
gotenberg-api = { git = "https://github.com/devmanorg/gotenberg-api.git" }
